+++ site/static/js/main.js
@@ -2,6 +2,7 @@ import './modernizr-custom.js';
 import './can-be-watched-button.component.js';
 import './external-purchase-button.component.js';
 import './carousel-video-mute-button.component.js';
+import './detail-player/detail-player.component.js';
 
 /*global Swiper, Modernizr, s72*/
 
@@ -566,4 +567,15 @@ function isTouchscreenEnabled() {
 document.addEventListener('s72loaded', event => {
   let app = event.detail.app;
   documentReady(app);
+});
+
+window.addEventListener('message', e => {
+  const {event, value} = e.data;
+  if (event == 's72-player:theatre-mode-change') {
+    const container = document.querySelector('#main .detail-player-container');
+    container.querySelector('detail-player').scrollIntoView({ behavior: 'smooth', block: 'start' });
+    container.classList.toggle('detail-player-theatre-mode', value);
+    document.querySelector('.meta-detail-bg').classList.toggle('meta-detail-bg--lights-out', value);
+    document.querySelector('.poster-wrapper').classList.toggle('d-none', value);
+  }
 });